+++ src/main/java/codesquad/webserver/WebServer.java
@@ -3,8 +3,8 @@ package codesquad.webserver;
 import codesquad.webserver.annotation.Autowired;
 import codesquad.webserver.annotation.Component;
 import codesquad.webserver.dispatcher.DispatcherServlet;
-import codesquad.webserver.dispatcher.StaticResourceHandler;
-import codesquad.webserver.dispatcher.StaticResourceResolver;
+import codesquad.webserver.staticresouce.StaticResourceHandler;
+import codesquad.webserver.staticresouce.StaticResourceResolver;
 import codesquad.webserver.httprequest.HttpRequest;
 import codesquad.webserver.httpresponse.HttpResponse;
 import codesquad.webserver.parser.HttpParser;
