--- src/main/java/codesquad/webserver/dispatcher/ViewResolver.java
@@ -1,7 +1,0 @@
-package codesquad.webserver.dispatcher;
-
-import codesquad.webserver.httprequest.HttpRequest;
-
-public interface ViewResolver {
-    View resolveView(ModelAndView modelAndView, HttpRequest request);
-}
