+++ src/main/java/codesquad/webserver/dispatcher/handler/adater/HandlerAdapter.java
@@ -1,5 +1,6 @@
-package codesquad.webserver.dispatcher;
+package codesquad.webserver.dispatcher.handler.adater;
 
+import codesquad.webserver.dispatcher.view.ModelAndView;
 import codesquad.webserver.httprequest.HttpRequest;
 
 public interface HandlerAdapter {
