+++ src/main/java/codesquad/webserver/dispatcher/requesthandler/HomeRequestHandler.java
@@ -1,4 +1,4 @@
-package codesquad.webserver.requesthandler;
+package codesquad.webserver.dispatcher.requesthandler;
 
 import static codesquad.webserver.httpresponse.HttpResponseBuilder.build;
 import static codesquad.webserver.httpresponse.HttpResponseBuilder.buildNotFoundResponse;
@@ -6,20 +6,20 @@ import static codesquad.webserver.httpresponse.HttpResponseBuilder.buildNotFound
 import codesquad.webserver.filereader.FileReader;
 import codesquad.webserver.httprequest.HttpRequest;
 import codesquad.webserver.httpresponse.HttpResponse;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 
-public class StaticFileHandler extends AbstractRequestHandler {
+public class HomeRequestHandler extends AbstractRequestHandler {
 
-    public StaticFileHandler(FileReader fileReader) {
+    private static final String FILE_PATH = "/index.html";
+
+    public HomeRequestHandler(FileReader fileReader) {
         super(fileReader);
     }
 
     @Override
     protected HttpResponse handleGet(HttpRequest request) {
         try {
-            FileReader.FileResource file = fileReader.read(request.requestLine().path());
+            FileReader.FileResource file = fileReader.read(FILE_PATH);
             return build(file);
         } catch (IOException e) {
             return buildNotFoundResponse();
