+++ src/main/java/codesquad/webserver/dispatcher/requesthandler/StaticFileHandler.java
@@ -1,4 +1,4 @@
-package codesquad.webserver.requesthandler;
+package codesquad.webserver.dispatcher.requesthandler;
 
 import static codesquad.webserver.httpresponse.HttpResponseBuilder.build;
 import static codesquad.webserver.httpresponse.HttpResponseBuilder.buildNotFoundResponse;
@@ -8,18 +8,16 @@ import codesquad.webserver.httprequest.HttpRequest;
 import codesquad.webserver.httpresponse.HttpResponse;
 import java.io.IOException;
 
-public class RegisterRequestHandler extends AbstractRequestHandler {
+public class StaticFileHandler extends AbstractRequestHandler {
 
-    private static final String FILE_PATH = "/registration/index.html";
-
-    public RegisterRequestHandler(FileReader fileReader) {
+    public StaticFileHandler(FileReader fileReader) {
         super(fileReader);
     }
 
     @Override
     protected HttpResponse handleGet(HttpRequest request) {
         try {
-            FileReader.FileResource file = fileReader.read(FILE_PATH);
+            FileReader.FileResource file = fileReader.read(request.requestLine().path());
             return build(file);
         } catch (IOException e) {
             return buildNotFoundResponse();
