+++ src/main/java/codesquad/webserver/dispatcher/view/JsonView.java
@@ -1,11 +1,10 @@
-package codesquad.webserver.dispatcher;
+package codesquad.webserver.dispatcher.view;
 
 import codesquad.webserver.httpresponse.HttpResponse;
 import java.util.Map;
 
-
-public class TemplateView implements View {
-    public TemplateView(String viewName, Map<String, Object> model) {
+public class JsonView implements View {
+    public JsonView(Map<String, Object> model) {
     }
 
     @Override
