+++ src/main/java/codesquad/webserver/dispatcher/view/View.java
@@ -1,4 +1,4 @@
-package codesquad.webserver.dispatcher;
+package codesquad.webserver.dispatcher.view;
 
 import codesquad.webserver.httpresponse.HttpResponse;
 import java.util.Map;
