+++ src/main/java/codesquad/webserver/dispatcher/view/resolver/ViewResolver.java
@@ -1,0 +1,9 @@
+package codesquad.webserver.dispatcher.view.resolver;
+
+import codesquad.webserver.dispatcher.view.ModelAndView;
+import codesquad.webserver.dispatcher.view.View;
+import codesquad.webserver.httprequest.HttpRequest;
+
+public interface ViewResolver {
+    View resolveView(ModelAndView modelAndView, HttpRequest request);
+}
